% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comments_clean.R
\name{comments_clean}
\alias{comments_clean}
\title{Clean comments}
\usage{
comments_clean(data)
}
\arguments{
\item{data}{created from using function \code{\link{recode_scores}}}
}
\value{
data frame
}
\description{
specifically for the comments in Best. Removes any records with
hardcoded 'N/A' and any blank rows. Ensures only unique comments are retained
and recodes emoticons/emojis (only those that have been added to function).
}
