% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_ch_mh.R
\name{filter_ch_mh}
\alias{filter_ch_mh}
\title{Filter data by CH or MH}
\usage{
filter_ch_mh(data, type = c("ch", "mh"))
}
\arguments{
\item{data}{data frame called responses}

\item{type}{String, either "ch" for community health or "mh" for mental
health. Will default to the first if not determined so will be "ch".}
}
\value{
data frame
}
\description{
Score counts are required for CH (Community Health) and MH
(Mental Health) but have different categories. This function does the counts
according to whether its ch or mh.
}
