% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_recorded.R
\name{filter_recorded}
\alias{filter_recorded}
\title{Filtering by recorded type}
\usage{
filter_recorded(data, type = c("ch", "mh"))
}
\arguments{
\item{data}{data frame called df}

\item{type}{String, either "ch" for community health or "mh" for mental
health. Will default to the first if not determined so will be "ch". Includes
function \code{\link{filter_ch_mh}} to separate data accordingly.}
}
\value{
data frame
}
\description{
Filtering by recorded type
}
