% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recode_scores.R
\name{recode_scores}
\alias{recode_scores}
\title{Clean Scores}
\usage{
recode_scores(data)
}
\arguments{
\item{data}{data frame for Scores and Scores2 data}
}
\value{
data frame
}
\description{
function to clean SQL code from Scores and Scores2 code, makes
SQL column names snake_case, select only necessary columns and recode the
long questions to match the MySQL column names of Listening, InvCare,
Service, Therapist and Best. Also filters out all questions
that are NA as a result of recoding.
}
