% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_files.R
\name{save_files}
\alias{save_files}
\title{Saving function}
\usage{
save_files(file, name_of_file)
}
\arguments{
\item{file}{data frame object created using \code{\link{filter_recorded}}}

\item{name_of_file}{String, suggest to make this the same as the object, will
be saved in the last month's folder as a .csv file.}
}
\value{
csv file
}
\description{
Saving in last month's newly created folder
}
