% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scores_cleaned.R
\name{scores_cleaned}
\alias{scores_cleaned}
\title{Cleaning Scores}
\usage{
scores_cleaned(data)
}
\arguments{
\item{data}{data frame for scores_old and scores_new data derived from the
function \code{\link{recode_scores}}.}
}
\value{
data frame
}
\description{
Cleaning Scores
}
